==> 03_appendix.Rmd <==
---
title: "Appendix analysis"
author: "Andrew Heiss"
date: "Last run: `r format(Sys.time(), '%B %e, %Y')`"
output:
  html_document:
    code_folding: hide
editor_options:
  chunk_output_type: console
---

```{r setup, message=FALSE}
knitr::opts_chunk$set(cache = FALSE, fig.retina = 2,
                      tidy.opts = list(width.cutoff = 120),  # For code
                      width = 120)  # For output
```

```{r load-libraries-data-functions, message=FALSE, warning=FALSE}
library(tidyverse)
library(ggstance)
library(broom)
library(countrycode)
library(pander)
library(stargazer)
library(skimr)
library(here)

# By default, R uses polynomial contrasts for ordered factors in linear models
# options("contrasts") 
# So make ordered factors use treatment contrasts instead
options(contrasts = rep("contr.treatment", 2))
# Or do it on a single variable:
# contrasts(df$x) <- "contr.treatment"

# Load pre-cleaned data
edb_clean <- read_rds(file.path(here(), "output", "data", "edb_clean.rds"))
edb_reforms <- read_rds(file.path(here(), "output", "data", "edb_reforms.rds"))

# Load helpful functions
source(file.path(here(), "lib", "model_stuff.R"))
source(file.path(here(), "lib", "graphics_stuff.R"))
```


# Variable descriptions

```{r tbl-vars-description, results="asis", warning=FALSE, message=FALSE}
edb_summary_base <- read_csv(file.path(here(), "data_raw", "variable-summary-base.csv"))

edb_var_descriptions <- edb_summary_base %>%
  mutate(Variable = paste0("`", Variable, "`"))

caption <- "Description of variables used in analysis"

tbl_var_descriptions <- pandoc.table.return(edb_var_descriptions, 
                                            caption = caption, 
                                            split.tables = Inf,
                                            justify = "lll", missing = "")

cat(tbl_var_descriptions)
cat(tbl_var_descriptions, 
    file = file.path(here(), "output", "tables", "app_table_var_desc.md"))
```


# Variable summaries

The small inline histograms in this table will only display correctly when they use a font that supports block elements, such as [DejaVu Sans](https://dejavu-fonts.github.io/) or Arial.

```{r tbl-vars-summary, results="asis", warning=FALSE, message=FALSE}
edb_summary_stats <- edb_clean %>%
  select(one_of(edb_summary_base$Variable)) %>%
  gather(Variable, value) %>%
  filter(!is.na(value)) %>%
  group_by(Variable) %>%
  summarize(N = n(),
            Mean = mean(value),
            `Std. Dev` = sd(value),
            Min = min(value),
            Max = max(value),
            Distribution = inline_hist(value)) 

edb_summary <- edb_summary_base %>%
  left_join(edb_summary_stats, by = "Variable") %>%
  mutate(Variable = paste0("`", Variable, "`")) %>%
  select(-Definition, -Source)

caption <- "Summary statistics of variables used in analysis"

tbl_edb_summary <- pandoc.table.return(edb_summary, caption = caption, 
                                       big.mark = ",", split.tables = Inf,
                                       justify = "lcccccc")

cat(tbl_edb_summary)
cat(tbl_edb_summary, 
    file = file.path(here(), "output", "tables", "app_table_var_summary.md"))
```


# List of countries in initial 2001 report

\* indicates country has an EDB reform committee by 2015

```{r tbl-countries, warning=FALSE, message=FALSE, results="asis"}
edb_bureaus <- read_csv(file.path(here(), "output", "data", "edb_bureaus.csv"))

country_names <- edb_clean %>%
  filter(in_2001 == 1) %>%
  group_by(ccode) %>% 
  summarize(Country = first(country_name)) %>% 
  ungroup() %>% 
  mutate(has_committee = ifelse(ccode %in% edb_bureaus$cowcode, "\\*", "")) %>%
  arrange(Country) %>%
  mutate(Country = paste0(Country, has_committee)) %>%
  select(Country)

caption <- "Countries in 2001 report"

tbl_countries <- pandoc.table.return(matrix(c(country_names$Country, rep(NA, 2)), 
                                            ncol = 4), caption = caption,
                                     split.tables = Inf, missing = "", justify = "llll")

cat(tbl_countries)
cat(tbl_countries, 
    file = file.path(here(), "output", "tables", "app_table_countries.md"))
```


# How control variables relate to being in the sample and to the outcomes

## Model A1

```{r model-in-2001, results="hide"}
edb_in_2001 <- edb_clean %>%
  filter(year > 2000) %>%
  group_by(ccode) %>%
  mutate(in_2001_lead = lead(in_2001),
         in_2001_lead = factor(in_2001_lead, levels = 0:1, 
                               labels = c("Not in 2001", "In 2001")))

model_in_2001 <- glm(in_2001_lead ~ gdpcap_ln + gdpgrowth + polity + pop_ln + 
                        inttot + civtot + loan_ln, 
                      data = edb_in_2001, family = binomial(link = "logit")) 

model_in_2001_robust <- robust_clusterify(model_in_2001, edb_in_2001, "ccode") %>%
  magrittr::use_series(coefs) %>% tidy()

model_in_2001_out <- stargazer(model_in_2001, type = "html", dep.var.caption = "",
                               dep.var.labels = "in\\_2001\\_lead",
                               se = list(model_in_2001_robust$std.error),
                               notes = c("Logistic regression model. Robust standard errors clustered by country.", 
                                         "Models include countries present in the 2001 EDB report."))
```

```{r tbl-model-in-2001, results="asis"}
cat(paste(escape_stars(model_in_2001_out), collapse = "\n"), "\n")
cat(paste(model_in_2001_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_model_A1.html"))
```

## Models A2–A5

```{r model-check-sb, results="hide"}
# This function generates an R formula based on a given outcome and its lead. 
# Given "sb_days_ln", it will create and run "sb_days_ln_lead ~ sb_days_ln + ..."
run_leaded_ols_check <- function(outcome, df) {
  outcome_lead <- paste0(outcome, "_lead")
  
  form <- as.formula(paste0(outcome_lead, " ~ ", outcome, 
                          " + gdpcap_ln + gdpgrowth + polity + ", 
                          "pop_ln + inttot + civtot + loan_ln"))
  
  lm(form, data = df)
}

# Define all the models that need to be run
models_to_run <- data_frame(outcome = c("sb_days_ln", "sb_proced",
                                        "sb_cost_ln", "sb_capital_ln"),
                            grouping = 1)  # Temporary variable for joining

edb_checks <- edb_clean %>%
  filter(year > 2000, in_2001 == 1) %>%
  mutate(grouping = 1) %>%
  group_by(ccode) %>%
  mutate_at(vars(sb_days_ln, sb_proced, sb_cost_ln, sb_capital_ln),
            funs(lead = lead(.))) %>%
  group_by(grouping) %>%
  nest() %>%
  right_join(models_to_run, by = "grouping")

# Run all the models within the data frame
edb_checks_models <- edb_checks %>%
  mutate(model = pmap(.l = list(outcome, data), run_leaded_ols_check),
         # Add robust clustered SEs
         robust_se = pmap(.l = list(model, data, "ccode"), robust_clusterify),
         # Add a data frame of model parameters with correct SEs
         tidy_robust = robust_se %>% map(~ tidy(.$coef)),
         ses_only = tidy_robust %>% map(~ .$std.error))

edb_checks_models_out <- stargazer(edb_checks_models$model, type = "html", 
                                   dep.var.caption = "",
                                   se = edb_checks_models$ses_only,
                                   notes = c("OLS models. Robust standard errors clustered by country.",
                                             "Models include countries present in the 2001 EDB report."),
                                   keep.stat = c("n", "rsq", "adj.rsq"))
```

```{r tbl-model-check-sb, results="asis"}
cat(paste(escape_stars(edb_checks_models_out), collapse = "\n"), "\n")
cat(paste(edb_checks_models_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_models_A2-5.html"))
```


# Starting a Business indicators

*Not generated with this script.*


# Enforcing Contracts indicators

*Not generated with this script.*


# Policy reform: OLS models

```{r models-ols-all}
# Define all the models that need to be run
models_to_run <- expand.grid(year = 2003:2013,
                             outcome = c("sb_proced", "sb_days_ln",
                                         "sb_cost_ln", "sb_capital_ln"),
                             grouping = c("All countries", 
                                          "No EDB reform committee", 
                                          "Special EDB reform committee"),
                             stringsAsFactors = FALSE)

# As in the paper analysis, this function generates an R model formula based on
# the name of the dependent variable and the year provided. i.e., given
# "sb_proced" and "2005", it will create the formula "sb_proced ~ sb_proced_lag
# + ranked_2005" and run the model
run_lagged_ols_model <- function(outcome, year, df) {
  outcome_lag <- paste0(outcome, "_lag")
  year_variable <- paste0("ranked_", year)

  form <- as.formula(paste0(outcome, " ~ ", outcome_lag, " + ", year_variable))

  lm(form, data = df)
}

dfs_split <- edb_clean %>%
  filter(in_2001 == 1) %>%
  group_by(has_bureau) %>%
  nest() %>%
  mutate(has_bureau = as.character(has_bureau)) %>%
  rename(grouping = has_bureau)

df_all <- edb_clean %>%
  filter(in_2001 == 1) %>%
  mutate(grouping = "All countries") %>%
  group_by(grouping) %>%
  nest()

models_to_run_full <- bind_rows(dfs_split, df_all) %>%
  right_join(models_to_run, by = "grouping")

# Run all the models within the data frame
ols_models_lagged <- models_to_run_full %>%
  mutate(model = pmap(.l = list(outcome, year, data), run_lagged_ols_model),
         # Add robust clustered SEs
         robust_se = pmap(.l = list(model, data, "ccode"), robust_clusterify),
         # Add model summary statistics
         glance = model %>% map(glance),
         # Add a data frame of model parameters with correct SEs
         tidy_robust = robust_se %>% map(~ tidy(.$coef)),
         ses_only = tidy_robust %>% map(~ .$std.error))

# Extract the ranking coefficients from all models
ols_coefs <- ols_models_lagged %>%
  # Spread out the model results
  unnest(tidy_robust) %>%
  # Only look at the ranked* coefficients
  filter(str_detect(term, "ranked")) %>%
  # Clean up the estimates, labels, and add stars
  mutate(value = paste0(sprintf("%.3f", round(estimate, 3)), p_stars(p.value)),
         term = str_replace(term, "\\.\\d+TRUE", "")) %>%
  # Get rid of extra columns
  select(-c(estimate, std.error, statistic, p.value)) %>%
  spread(outcome, value) %>%
  # Give table clean column names
  select(Subset = grouping, Year = year,
         Procedures = sb_proced, `Cost (log)` = sb_cost_ln,
         `Days (log)` = sb_days_ln, `Capital (log)` = sb_capital_ln)
```


## Table 3 from the paper: summary of coefficients for 8 OLS models

Full models in tables B1–3 below.

```{r tbl-ols-in-paper, results="asis"}
tbl_ols_paper <- ols_coefs %>%
  filter(Subset == "All countries", Year %in% c(2005, 2006))

caption <- 'Summary of β~2~ coefficients (i.e. "ranked_200x") for difference models'

tbl_ols <- pandoc.table.return(tbl_ols_paper, caption = caption)

cat(tbl_ols)
cat(tbl_ols, file = file.path(here(), "output", "tables", "app_table_ols.md"))
```

## Table B1: 2005 analysis, full OLS models

```{r tbl-ols-2005-2006, results="hide"}
tbl_ols_2005 <- ols_models_lagged %>%
  filter(grouping == "All countries", year == 2005)

tbl_ols_2005_out <- stargazer(tbl_ols_2005$model, type = "html", 
                              dep.var.caption = "",
                              se = tbl_ols_2005$ses_only,
                              notes = c("OLS models. Robust standard errors clustered by country.", 
                                        "Models include countries present in the 2001 EDB report."),
                              keep.stat = c("n", "rsq", "adj.rsq"))

tbl_ols_2006 <- ols_models_lagged %>%
  filter(grouping == "All countries", year == 2006)

tbl_ols_2006_out <- stargazer(tbl_ols_2006$model, type = "html", 
                              dep.var.caption = "",
                              se = tbl_ols_2006$ses_only,
                              notes = c("OLS models. Robust standard errors clustered by country.", 
                                        "Models include countries present in the 2001 EDB report."),
                              keep.stat = c("n", "rsq", "adj.rsq"))
```

```{r tbl-ols-2005, results="asis"}
cat(paste(escape_stars(tbl_ols_2005_out), collapse = "\n"), "\n")
cat(paste(tbl_ols_2005_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_models_B1.html"))
```

## Table B2: 2006 analysis, full OLS models

```{r tbl-ols-2006, results="asis"}
cat(paste(escape_stars(tbl_ols_2006_out), collapse = "\n"), "\n")
cat(paste(tbl_ols_2006_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_models_B2.html"))
```

## Table B3: Cutpoints at every possible year, OLS models

```{r tbl-ols-all, results="asis"}
caption <- 'Summary of β~2~ coefficients (i.e. "ranked.200x") for difference models for all years'

tbl_ols_all <- pandoc.table.return(ols_coefs, caption = caption,
                                     split.tables = Inf, missing = "—")

cat(tbl_ols_all)
cat(tbl_ols_all, 
    file = file.path(here(), "output", "tables", "app_models_B3_all.md"))
```


# Policy reform: ITS models

```{r models-its-all}
run_its_model <- function(outcome, year, df) {
  year_variable <- paste0("ranked_", year)
  year_centered <- paste0("year_centered_", year)
  
  form <- as.formula(paste0(outcome, " ~ ", year_centered, " + ", year_variable,
                            " + ", year_centered, " * ", year_variable))
  
  lm(form, data = df)
}

# Run all the models within the data frame
its_models_lagged <- models_to_run_full %>%
  mutate(model = pmap(.l = list(outcome, year, data), run_its_model),
         # Add robust clustered SEs
         robust_se = pmap(.l = list(model, data, "ccode"), robust_clusterify),
         # Add model summary statistics
         glance = model %>% map(glance),
         # Add a data frame of model parameters with correct SEs
         tidy_robust = robust_se %>% map(~ tidy(.$coef)),
         ses_only = tidy_robust %>% map(~ .$std.error))

# Extract the ranking coefficients from all models
its_coefs <- its_models_lagged %>%
  # Spread out the model results
  unnest(tidy_robust) %>%
  # Only look at the coefficients from interaction terms (they have ":" in their names)
  filter(str_detect(term, ":")) %>%
  # Clean up the estimates, labels, and add stars
  mutate(value = paste0(sprintf("%.3f", round(estimate, 3)), p_stars(p.value)),
         term = str_replace(term, "(.+)\\.\\d+:(.+)\\.\\d+TRUE", "\\1 × \\2")) %>%
  # Get rid of extra columns
  select(-c(estimate, std.error, statistic, p.value)) %>%
  spread(outcome, value) %>%
  # Give table clean column names
  select(Subset = grouping, Year = year,
         Procedures = sb_proced, `Cost (log)` = sb_cost_ln,
         `Days (log)` = sb_days_ln, `Capital (log)` = sb_capital_ln)
```

## Table 4 from the paper: summary of coefficients for 8 ITS models

Full models in tables C1–3 below.

```{r tbl-its-in-paper, results="asis"}
tbl_its_paper <- its_coefs %>%
  filter(Subset == "All countries", Year %in% c(2005, 2006))

caption <- 'Summary of β~3~ coefficients (i.e. "year.centered.200x × ranked.200x") for ITS models'

tbl_its <- pandoc.table.return(tbl_its_paper, caption = caption)

cat(tbl_its)
cat(tbl_its, file = file.path(here(), "output", "tables", "app_table_its.md"))
```

## Table C1: 2005 analysis, full ITS models

```{r tbl-its-2005-2006, results="hide"}
tbl_its_2005 <- its_models_lagged %>%
  filter(grouping == "All countries", year == 2005)

tbl_its_2005_out <- stargazer(tbl_its_2005$model, type = "html", 
                              dep.var.caption = "",
                              se = tbl_its_2005$ses_only,
                              notes = c("ITS models. Robust standard errors clustered by country.", 
                                        "Models include countries present in the 2001 EDB report."),
                              keep.stat = c("n", "rsq", "adj.rsq"))

tbl_its_2006 <- its_models_lagged %>%
  filter(grouping == "All countries", year == 2006)

tbl_its_2006_out <- stargazer(tbl_its_2006$model, type = "html", 
                              dep.var.caption = "",
                              se = tbl_its_2006$ses_only,
                              notes = c("ITS models. Robust standard errors clustered by country.", 
                                        "Models include countries present in the 2001 EDB report."),
                              keep.stat = c("n", "rsq", "adj.rsq"))
```

```{r tbl-its-2005, results="asis"}
cat(paste(escape_stars(tbl_its_2005_out), collapse = "\n"), "\n")
cat(paste(tbl_its_2005_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_models_C1.html"))
```

## Table C2: 2006 analysis, full ITS models

```{r tbl-its-2006, results="asis"}
cat(paste(escape_stars(tbl_its_2006_out), collapse = "\n"), "\n")
cat(paste(tbl_its_2006_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_models_C2.html"))
```

## Table C3: Cutpoints at every possible year, ITS models

```{r tbl-its-all, results="asis"}
caption <- 'Summary of β~3~ coefficients (i.e. "year.centered.200x × ranked.200x") for ITS models for all years'

tbl_its_all <- pandoc.table.return(its_coefs, caption = caption,
                                   split.tables = Inf, missing = "—")

cat(tbl_its_all)
cat(tbl_its_all, 
    file = file.path(here(), "output", "tables", "app_models_C3_all.md"))
```


# Country fixed effects for OLS models

## Table D1: "Ranked" coefficients with and without country fixed effects

```{r ranked-country-fe, warning=FALSE}
# All the FE models to be run
country_fe_models <- expand.grid(outcome = c("sb_proced", "sb_days_ln",
                                             "sb_cost_ln", "sb_capital_ln",
                                             "con_proced", "con_days"),
                                 controls = c("Main", 
                                              "Main + `civtot_lag` + `inttot_lag` + `loan_ln_lag`", 
                                              "Main + `civtot_lag` + `inttot_lag` + `loan_bin_lag`", 
                                              "Main + fixed effects",
                                              "Main - 2001", "Main - 2001 & 2002"),
                                 stringsAsFactors = FALSE) %>%
  mutate(country_fe = controls %in% c("Main + fixed effects", "Main - 2001", "Main - 2001 & 2002"),
         df_to_use = case_when(
           controls == "Main - 2001" ~ "Drop 2001",
           controls == "Main - 2001 & 2002" ~ "Drop 2002",
           TRUE ~ "Full" 
         ))

# Nest filtered data into a data frame
data_to_use <- tribble(
  ~df_to_use,  ~data,
  "Full",      edb_clean %>% filter(year > 2000, in_2004 == 1),
  "Drop 2001", edb_clean %>% filter(year > 2001, in_2004 == 1),
  "Drop 2002", edb_clean %>% filter(year > 2002, in_2004 == 1)
)

# Combine filtered data with all model parameters
country_fe_models_full <- data_to_use %>%
  right_join(country_fe_models, by = "df_to_use") %>%
  select(-df_to_use)

# Build and run a bunch of different models based on parameters that determine
# what controls to add and whether or not to include country fixed effects
run_country_fe_models <- function(outcome, controls, country_fe, df) {
  main_controls <- "gdpcap_ln_lag + gdpgrowth_lag + polity_lag + pop_ln_lag"
  
  if (str_detect(controls, "loan_ln_lag")) {
    additional_controls <- " + civtot_lag + inttot_lag + loan_ln_lag"
  } else if (str_detect(controls, "loan_bin_lag")) {
    additional_controls <- " + civtot_lag + inttot_lag + loan_bin_lag"
  } else {
    additional_controls <- ""
  }
  
  if (country_fe) {
    country <- " + as.factor(ccode)"
  } else {
    country <- ""
  }

  form <- paste0(outcome, " ~ ", paste0(outcome, "_lag"), " + ranked_lag + ", 
                 main_controls, additional_controls, country) %>%
    as.formula()

  lm(form, data = df)
}

# Run all the models within the data frame
country_fe_models_all <- country_fe_models_full %>%
  mutate(model = pmap(.l = list(outcome, controls, country_fe, data), 
                      run_country_fe_models),
         # Add robust clustered SEs
         robust_se = pmap(.l = list(model, data, "ccode"), robust_clusterify),
         # Add model summary statistics
         glance = model %>% map(glance),
         # Add a data frame of model parameters with correct SEs
         tidy_robust = robust_se %>% map(~ tidy(.$coef)),
         ses_only = tidy_robust %>% map(~ .$std.error))

# Make tiny data frame indicating if model had country fixed effects
fixed_effects_country <- country_fe_models %>%
  select(controls, country_fe) %>%
  distinct() %>%
  mutate(country_fe = ifelse(country_fe, "Yes", "No")) %>%
  spread(controls, country_fe) %>%
  mutate(Outcome = "Fixed country effects")

# Display all the ranked coefficients
ranked_coefs_country <- country_fe_models_all %>%
  # Spread out the model results
  unnest(tidy_robust) %>%
  filter(str_detect(term, "ranked")) %>%
  # Clean up the estimates and add stars
  mutate(value = paste0(sprintf("%.3f", round(estimate, 3)), p_stars(p.value)),
         outcome = factor(outcome, levels = unique(country_fe_models$outcome),
                          labels = paste0("`", unique(country_fe_models$outcome), "`"),
                          ordered = TRUE),
         controls = factor(controls, levels = unique(country_fe_models$controls),
                           ordered = TRUE)) %>%
  # Get rid of extra columns
  select(Outcome = outcome, controls, value) %>%
  spread(controls, value) %>%
  bind_rows(fixed_effects_country)
```

```{r tbl-country-fe-ranked, results="asis"}
caption <- 'Summary of coefficients for the lagged "Ranked" variable with and without country fixed effects'

tbl_country_fe <- pandoc.table.return(ranked_coefs_country, caption = caption,
                                      split.tables = Inf)

cat(tbl_country_fe)
cat(tbl_country_fe, 
    file = file.path(here(), "output", "tables", "app_table_D1.md"))
```

*Note*: Main controls are `gdpcap_ln_lag`, `gdpgrowth_lag`, `polity_lag`, and `pop_ln_lag`


## Table D2: Sub-indicators for "Starting a Business" with and without country fixed effects

```{r sb-country-fe, results="hide"}
sb_fe <- country_fe_models_all %>%
  filter(str_detect(outcome, "sb_"),
         controls %in% c("Main + fixed effects", 
                         "Main + `civtot_lag` + `inttot_lag` + `loan_bin_lag`")) %>%
  mutate(outcome = fct_inorder(outcome, ordered = TRUE)) %>%
  arrange(outcome)

sb_country_fe_models_out <- stargazer(sb_fe$model, type = "html", 
                                      dep.var.caption = "",
                                      se = sb_fe$ses_only,
                                      notes = c("OLS models. Robust standard errors clustered by country.",
                                                "Models include countries present in the 2004 EDB report."),
                                      keep.stat = c("n", "rsq", "adj.rsq"),
                                      add.lines = list(c("Country fixed effects", 
                                                         rep(c("No", "Yes"), 4))),
                                      omit = "ccode")
```

```{r tbl-sb-country-fe, results="asis"}
cat(paste(escape_stars(sb_country_fe_models_out), collapse = "\n"), "\n")
cat(paste(sb_country_fe_models_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_table_D2.html"))
```

## Table D3: Sub-indicators for "Enforcing Contracts" with and without country fixed effects

```{r con-country-fe, results="hide"}
con_fe <- country_fe_models_all %>%
  filter(str_detect(outcome, "con_"),
         controls %in% c("Main + fixed effects", 
                         "Main + `civtot_lag` + `inttot_lag` + `loan_bin_lag`")) %>%
  mutate(outcome = fct_inorder(outcome, ordered = TRUE)) %>%
  arrange(outcome)

con_country_fe_models_out <- stargazer(con_fe$model, type = "html", 
                                       dep.var.caption = "",
                                       se = con_fe$ses_only,
                                       notes = c("OLS models. Robust standard errors clustered by country.",
                                                 "Models include countries present in the 2004 EDB report."),
                                       keep.stat = c("n", "rsq", "adj.rsq"),
                                       add.lines = list(c("Country fixed effects", 
                                                          rep(c("No", "Yes"), 4))),
                                       omit = "ccode")
```

```{r tbl-con-country-fe, results="asis"}
cat(paste(escape_stars(con_country_fe_models_out), collapse = "\n"), "\n")
cat(paste(con_country_fe_models_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_table_D3.html"))
```


# Year fixed effects for OLS models

## Table E1: "Ranked" coefficients with and without year fixed effects

```{r ranked-year-fe, warning=FALSE}
# All the FE models to be run
year_fe_models <- expand.grid(outcome = c("sb_proced", "sb_days_ln",
                                          "sb_cost_ln", "sb_capital_ln",
                                          "con_proced", "con_days"),
                              controls = c("Main", 
                                           "Main + `civtot_lag` + `inttot_lag` + `loan_ln_lag`", 
                                           "Main + `civtot_lag` + `inttot_lag` + `loan_bin_lag`", 
                                           "Main + fixed effects",
                                           "Main - 2001", "Main - 2001 & 2002"),
                              stringsAsFactors = FALSE) %>%
  mutate(year_fe = controls %in% c("Main + fixed effects", "Main - 2001", "Main - 2001 & 2002"),
         df_to_use = case_when(
           controls == "Main - 2001" ~ "Drop 2001",
           controls == "Main - 2001 & 2002" ~ "Drop 2002",
           TRUE ~ "Full" 
         ))

# Nest filtered data into a data frame
data_to_use <- tribble(
  ~df_to_use,  ~data,
  "Full",      edb_clean %>% filter(year > 2000, in_2004 == 1),
  "Drop 2001", edb_clean %>% filter(year > 2001, in_2004 == 1),
  "Drop 2002", edb_clean %>% filter(year > 2002, in_2004 == 1)
)

# Combine filtered data with all model parameters
year_fe_models_full <- data_to_use %>%
  right_join(year_fe_models, by = "df_to_use") %>%
  select(-df_to_use)

# Build and run a bunch of different models based on parameters that determine
# what controls to add and whether or not to include country fixed effects
run_year_fe_models <- function(outcome, controls, year_fe, df) {
  main_controls <- "gdpcap_ln_lag + gdpgrowth_lag + polity_lag + pop_ln_lag"
  
  if (str_detect(controls, "loan_ln_lag")) {
    additional_controls <- " + civtot_lag + inttot_lag + loan_ln_lag"
  } else if (str_detect(controls, "loan_bin_lag")) {
    additional_controls <- " + civtot_lag + inttot_lag + loan_bin_lag"
  } else {
    additional_controls <- ""
  }
  
  if (year_fe) {
    year <- " + as.factor(year)"
  } else {
    year <- ""
  }

  form <- paste0(outcome, " ~ ", paste0(outcome, "_lag"), " + ranked_lag + ", 
                 main_controls, additional_controls, year) %>%
    as.formula()

  lm(form, data = df)
}

# Run all the models within the data frame
year_fe_models_all <- year_fe_models_full %>%
  mutate(model = pmap(.l = list(outcome, controls, year_fe, data), 
                      run_year_fe_models),
         # Add robust clustered SEs
         robust_se = pmap(.l = list(model, data, "ccode"), robust_clusterify),
         # Add model summary statistics
         glance = model %>% map(glance),
         # Add a data frame of model parameters with correct SEs
         tidy_robust = robust_se %>% map(~ tidy(.$coef)),
         ses_only = tidy_robust %>% map(~ .$std.error))

# Make tiny data frame indicating if model had country fixed effects
fixed_effects_year <- year_fe_models %>%
  select(controls, year_fe) %>%
  distinct() %>%
  mutate(year_fe = ifelse(year_fe, "Yes", "No")) %>%
  spread(controls, year_fe) %>%
  mutate(Outcome = "Fixed year effects")

# Display all the ranked coefficients
ranked_coefs_year <- year_fe_models_all %>%
  # Spread out the model results
  unnest(tidy_robust) %>%
  filter(str_detect(term, "ranked")) %>%
  # Clean up the estimates and add stars
  mutate(value = paste0(sprintf("%.3f", round(estimate, 3)), p_stars(p.value)),
         outcome = factor(outcome, levels = unique(year_fe_models$outcome),
                          labels = paste0("`", unique(year_fe_models$outcome), "`"),
                          ordered = TRUE),
         controls = factor(controls, levels = unique(year_fe_models$controls),
                           ordered = TRUE)) %>%
  # Get rid of extra columns
  select(Outcome = outcome, controls, value) %>%
  spread(controls, value) %>%
  bind_rows(fixed_effects_year)
```

```{r tbl-year-fe-ranked, results="asis"}
caption <- 'Summary of coefficients for the lagged "Ranked" variable with and without country fixed effects'

tbl_year_fe <- pandoc.table.return(ranked_coefs_year, caption = caption,
                                   split.tables = Inf)

cat(tbl_year_fe)
cat(tbl_year_fe, 
    file = file.path(here(), "output", "tables", "app_table_E1.md"))
```

*Note*: Main controls are `gdpcap_ln_lag`, `gdpgrowth_lag`, `polity_lag`, and `pop_ln_lag`

## Table E2: Sub-indicators for "Starting a Business" with and without year fixed effects

```{r sb-year-fe, results="hide"}
sb_fe_year <- year_fe_models_all %>%
  filter(str_detect(outcome, "sb_"),
         controls %in% c("Main + fixed effects", 
                         "Main + `civtot_lag` + `inttot_lag` + `loan_bin_lag`")) %>%
  mutate(outcome = fct_inorder(outcome, ordered = TRUE)) %>%
  arrange(outcome)

sb_year_fe_models_out <- stargazer(sb_fe_year$model, type = "html", 
                                   dep.var.caption = "",
                                   se = sb_fe_year$ses_only,
                                   notes = c("OLS models. Robust standard errors clustered by country.",
                                             "Models include countries present in the 2004 EDB report."),
                                   keep.stat = c("n", "rsq", "adj.rsq"),
                                   add.lines = list(c("Year fixed effects", 
                                                      rep(c("No", "Yes"), 4))),
                                   omit = "year")
```

```{r tbl-sb-year-fe, results="asis"}
cat(paste(escape_stars(sb_year_fe_models_out), collapse = "\n"), "\n")
cat(paste(sb_year_fe_models_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_table_E2.html"))
```

## Table E3: Sub-indicators for "Enforcing Contracts" with and without year fixed effects

```{r con-year-fe, results="hide"}
con_fe_year <- year_fe_models_all %>%
  filter(str_detect(outcome, "con_"),
         controls %in% c("Main + fixed effects", 
                         "Main + `civtot_lag` + `inttot_lag` + `loan_bin_lag`")) %>%
  mutate(outcome = fct_inorder(outcome, ordered = TRUE)) %>%
  arrange(outcome)

con_year_fe_models_out <- stargazer(con_fe_year$model, type = "html", 
                                    dep.var.caption = "",
                                    se = con_fe_year$ses_only,
                                    notes = c("OLS models. Robust standard errors clustered by country.",
                                              "Models include countries present in the 2004 EDB report."),
                                    keep.stat = c("n", "rsq", "adj.rsq"),
                                    add.lines = list(c("Year fixed effects", 
                                                       rep(c("No", "Yes"), 4))),
                                    omit = "year")
```

```{r tbl-con-year-fe, results="asis"}
cat(paste(escape_stars(con_year_fe_models_out), collapse = "\n"), "\n")
cat(paste(con_year_fe_models_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_table_E3.html"))
```


# Description of Reform Coding

*Not generated with this script.*


# What explains reform committees?

## Table F1: Correlates of reform committee presence

```{r reform-committee-presence, results="hide"}
model_committee = glm(has_bureau ~ p_edb_rank + icrg_index + 
                        gdpcap + gdpgrowth + fdi_inper + trade + log1p(ibrd) +
                        polity + yrsoffc,
                      data = filter(edb_clean, year == 2008),
                      family = binomial(link = "logit"))

model_committee_robust <- robust_clusterify(model_committee, 
                                            filter(edb_clean, year == 2008), "ccode") %>%
  magrittr::use_series(coefs) %>% tidy()

model_committee_out <- stargazer(model_committee, type = "html", 
                                 dep.var.caption = "",
                                 dep.var.labels = "has\\_bureau",
                                 se = list(model_committee_robust$std.error),
                                 notes = "Logsitic regression model. Year restricted to 2008. Robust standard errors clustered by country.")
```

```{r tbl-reform-committee-presence, results="asis"}
cat(paste(escape_stars(model_committee_out), collapse = "\n"), "\n")
cat(paste(model_committee_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_table_F1.html"))
```

## Figure F1: Correlates of reform committee presence

Coefficients from Table F1

```{r fig-reform-committee-presence, fig.width=6, fig.height=4}
committee_tidy_robust <- model_committee_robust %>%
  filter(term != "(Intercept)") %>%
  mutate(conf.low = estimate - std.error * qnorm(0.975),
         conf.high = estimate + std.error * qnorm(0.975)) %>%
  mutate_at(vars(estimate, conf.low, conf.high),
            funs(exp = exp(.))) %>%
  mutate(term = fct_inorder(term, ordered = TRUE))

plot_model_committee <- ggplot(committee_tidy_robust, 
                               aes(x = estimate_exp, y = fct_rev(term))) +
  geom_vline(xintercept = 1, colour = "black") +
  geom_pointrangeh(aes(xmin = conf.low_exp, xmax = conf.high_exp), size = 0.5) + 
  labs(x = "Odds ratio", y = NULL) +
  theme_edb()
plot_model_committee

ggsave(plot_model_committee, 
       filename = file.path(here(), "output", "figures", "app_committee_presence.pdf"),
       width = 6, height = 4, units = "in", device = cairo_pdf)
ggsave(plot_model_committee, 
       filename = file.path(here(), "output", "figures", "app_committee_presence.png"),
       width = 6, height = 4, units = "in", type = "cairo", dpi = 300)
```

## Table F2: Correlates of control variables from Models A1–A5

```{r tbl-sig-summary-a1-a5, results="asis"}
edb_checks_a2_a5 <- edb_checks_models %>%
  unnest(tidy_robust) %>% select(-grouping)

edb_checks_a1 <- model_in_2001_robust %>% mutate(outcome = "in_2001")

model_names <- tribble(
  ~model, ~outcome,
  "A",    "in_2001",
  "B",    "sb_days_ln",
  "C",    "sb_proced",
  "D",    "sb_cost_ln",
  "E",    "sb_capital_ln"
) %>%
  mutate(name_clean = paste0("Model ", model, "<br>", "`", outcome, "`"))

edb_checks_significance <- bind_rows(edb_checks_a1, edb_checks_a2_a5) %>%
  left_join(model_names, by = "outcome") %>%
  filter(term != "(Intercept)", !str_detect(term, "sb_")) %>%
  mutate(significance = case_when(
    p.value < 0.10 & sign(estimate) == 1 ~ "+",
    p.value < 0.10 & sign(estimate) == -1 ~ "−",
    TRUE ~ ""
  )) %>%
  mutate(term = paste0("`", term, "`"),
         term = fct_inorder(term, ordered = TRUE)) %>%
  select(` ` = term, significance, name_clean) %>%
  spread(name_clean, significance)

caption <- "Correlates of relevant variables with selection into 2001 sample and with outcomes"

tbl_edb_checks_sig <- pandoc.table.return(edb_checks_significance, 
                                          caption = caption, 
                                          split.tables = Inf,
                                          justify = "lccccc", missing = "", 
                                          style = "multiline",
                                          keep.line.breaks = TRUE)

cat(tbl_edb_checks_sig)
cat(tbl_edb_checks_sig, 
    file = file.path(here(), "output", "tables", "app_table_F2.md"))
```

*Note*: A negative coefficient for the outcome variables indicates improvements.

For models and measures, see section "How control variables relate to being in the sample and to the outcomes" above.


## Table F3: Rankings in 2005 and 2014

```{r reform-committee-2005-2014, results="hide"}
edb_rankings <- edb_clean %>%
  select(ccode, year, p_edb_rank, has_bureau) %>%
  filter(year %in% c(2005, 2014)) %>%
  spread(year, p_edb_rank) %>%
  mutate(change_in_ranking = `2014` - `2005`)

model_rankings <- lm(`2014` ~ `2005` + has_bureau,
                     data = edb_rankings)

model_rankings_interaction <- lm(`2014` ~ `2005` + has_bureau + `2005` * has_bureau,
                                 data = edb_rankings)

model_rankings_change <- lm(change_in_ranking ~ has_bureau,
                            data = edb_rankings)

var_labs <- c("2005 ranking", "Has reform committee",
              "2005 ranking × has reform committee")

model_rankings_out <- stargazer(model_rankings, model_rankings_interaction, 
                                model_rankings_change,
                                type = "html",
                                dep.var.caption = "",
                                dep.var.labels = c("2014", "Change in ranking"),
                                covariate.labels = var_labs,
                                notes = "OLS models. Non-robust standard errors.",
                                keep.stat = c("n", "rsq", "adj.rsq"))
```

```{r tbl-reform-committee-2005-2014, results="asis"}
cat(paste(escape_stars(model_rankings_out), collapse = "\n"), "\n")
cat(paste(model_rankings_out, collapse = "\n"), "\n",
    file = file.path(here(), "output", "tables", "app_table_F3.html"))
```


# India experiment survey text

*Not generated with this script.*


# Investor experiment survey text

*Not generated with this script.*
